requirements.txt
streamlit==1.39.0
google-generativeai==0.7.2
fpdf2==2.7.9
python-dotenv==1.0.1
